tech-debt: extract seed-testing-availability.helpers (ROK-1183)
